refactor(sponsor-reports): co-locate report-errors + reports-money out of shared utils
Both files moved into src/pages/sponsors/sponsor-reports/ per §2 (feature helpers
never in shared utils/); test for reports-money co-located alongside. Internal
./constants import in report-errors updated to ../../../utils/constants.
formatUsd logic unchanged — uicore/cents contract deferred to PR-B.
