+++ src/components/sponsors/reports/OrdersTable.js
@@ -14,7 +14,7 @@
 import React from "react";
 import MuiTable from "openstack-uicore-foundation/lib/components/mui/table";
 import StatusPill from "./StatusPill";
-import { formatUsd } from "../../../utils/reports-money";
+import { formatUsd } from "../../../pages/sponsors/sponsor-reports/reports-money";
 
 const ISO_DATE_LENGTH = 10; // "YYYY-MM-DD"
 const MS_PER_SECOND = 1000;
